UPSTREAM: <carry>: kube-controller-manager: allow running bare kube-controller-manager
